% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_reg}
\alias{A_reg}
\title{Albanese et al. (2016) regression data}
\format{
A data frame with 4 rows and 3 columns where
the rows are for maximum femur, tibia, humerus, and radius lengths
and the columns are the regression coefficient, the regression
constant, and the standard error of estimate. The bone measurements
should be in mm. and the returned stature estimates are in mm.
}
\usage{
data(A_reg)
}
\description{
Contains the bivariate regression data for stature on long bones from
Albanese et al. (2016) Table 3.
}
\references{
Albanese, J., Tuck, A., Gomes, J., & Cardoso, H. F. (2016). An alternative
approach for estimating stature from long bones that is not population-or
group-specific. \emph{Forensic Science International}, 259, 59-68.
}
\keyword{datasets}
