% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Europe2}
\alias{Europe2}
\title{Subset of European data}
\format{
A data frame with 536 rows and 11 variables
}
\source{
{
See "European data set: Excel" and "Notes" under "Special Projects"
\url{https://fae.johnshopkins.edu/chris-ruff/}
}
}
\usage{
data(Europe2)
}
\description{
A subset of the European data.  The data come from Ruff (2018).  Excludes individuals
with statures that were estimated using Niskanen & Ruff (2018).
}
\details{
Data as follows:
\itemize{
 \item{Region:}{  Geographic region}
 \item{Subregion:}{  Geographic subregion}
 \item{Site:}{  Site}
 \item{ID:}{  Unique identifier}
 \item{Museum:}{  Museum}
 \item{Sex:}{  Sex as a factor, either "Female" or "Male"}
 \item{State:}{  Stature in mm.}
 \item{Fem:}{  Maximum femur length in mm.}
 \item{Tib:}{  Maximum tibia length in mm.}
 \item{Hum:}{  Maximum humerus length in mm.}
 \item{Rad:}{  Maximum radius length in mm.}
}
}
\references{
Raxter, M.H., Auerbach, B.M. and Ruff, C.B. (2006) A revision of the Fully technique
for estimating statures. \emph{Am. J. Phys. Anthropol}. 130, 374-384.

Raxter, M.H., Ruff, C.B. and Auerbach, B.M. (2007) Technical note: Revised Fully
stature estimation technique. \emph{Am. J. Phys. Anthropol}. 133, 817-818.

Ruff, C.B. (ed.) (2018) \emph{Skeletal variation and adaptation in Europeans:
Upper Paleolithic to the Twentieth Century}. Hoboken: Wiley Blackwell.
}
\keyword{datasets}
