% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig11_07.R
\name{Fig11_07}
\alias{Fig11_07}
\title{Preece-Baines growth (distance) curve fit to longitudinal growth data}
\usage{
Fig11_07()
}
\value{
Plot of data points and Preece-Baines curve.
}
\description{
Fits a Preece-Baines growth curve to data given in Scammon (1927).
}
\details{
The object is called Fig11_07 as it is similar to Figure 11.7 on page 271 of Mielke, Konigsberg,
and Relethford (2011) and on page 261 of the 2006 edition where it is Figure 10.7.
It plots the data points from Scammon (1927) and the Preece-Baines (1978) curve
fit to the data points.  This script is "hard-wired," so there are no calling
parameters.
}
\references{
Mielke, James H., Konigsberg, Lyle W., and Relethford, John H. (2011)
\emph{Human biological variation}, 2nd ed.  New York: Oxford University Press.

Mielke, James H., Konigsberg, Lyle W., and Relethford, John H. (2006)
\emph{Human biological variation}, 1st ed.  New York: Oxford University Press.

Preece, M. A., & Baines, M. J. (1978). A new family of mathematical models
describing the human growth curve. \emph{Annals of Human Biology}, 5, 1-24.

Scammon, R. E. (1927). The first seriatim study of human growth.
\emph{American Journal of Physical Anthropology}, 10(3), 329-336.
}
