% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig3_13.R
\name{Fig3_13}
\alias{Fig3_13}
\title{Plot of average relative fitness against allele frequencies.}
\usage{
Fig3_13(
  w = c(0.7635, 1, 1e-04),
  x = c(0.65, 0.95),
  y = c(0.77, 0.82),
  tan.int = 0.01
)
}
\arguments{
\item{w}{Vector of three values giving the relative fitness for each genotype,
AA, AB, and BB.  As these are relative fitness values, at least one value must
be 1.0. Default = c(0.7635,1,0.0001)}

\item{x}{A vector containing two elements that give the limits for the allele
frequencies for the plot.  This is used to "zoom in" on the graph.  Default=c(0.65,0.95)}

\item{y}{A vector containing two elements that give the limits for the relative
fitness values for the plot.  This is used to "zoom in" on the graph.  Default=c(0.77,0.82)}

\item{tan.int}{A real value giving the frequency of the tangents to the curve.
The constraint is that 0<= tan.int <=1. Values closer to 0.0 give more tangents
and  values closer to 1 give fewer tangents. Default=0.01}
}
\value{
Plot of average relative fitness against allele frequencies with tangents
drawn to the curve.
}
\description{
Shows the curve of average relative fitness against allele frequencies
in a biallelic system under selection regimes.
}
\details{
The object is called Fig3_13 as it is Figure 3.13 on page 74 of Mielke, Konigsberg,
and Relethford (2011) and on page 70 of the 2006 edition where it is Figure 3.8.
}
\references{
Mielke, James H., Konigsberg, Lyle W., and Relethford, John H. (2011)
\emph{Human biological variation}, 2nd ed.  New York: Oxford University Press.

Mielke, James H., Konigsberg, Lyle W., and Relethford, John H. (2006)
\emph{Human biological variation}, 1st ed.  New York: Oxford University Press.
}
