% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig9_6.R
\name{Fig9_6}
\alias{Fig9_6}
\title{Plot of two coalescent trees}
\usage{
Fig9_6(popsize1 = 100, popsize2 = 1000)
}
\arguments{
\item{popsize1}{Integer value giving population size for first coalescent tree.
Default = 100}

\item{popsize2}{Integer value giving population size for first coalescent tree.
Default = 1000}
}
\value{
Plot of two coalescent trees.
}
\description{
Plots simulation of two coalescent trees.
}
\details{
The object is called Fig9_6 as it is Figure 9.6 on page 212 of Mielke, Konigsberg,
and Relethford (2011) and on page 202 of the 2006 edition where it is Figure 8.6.
It plots two coalescent trees side by side (on the same scale) each with 20 "tips"
(current alleles). The two trees can represent different population sizes. The
point of this is to show the relationship between trees based on different population
sizes, where coalescent trees are shallower if the population size is smaller
and deeper if the population size is larger.  To wit, coalescence is more rapid
when the population size is small and slower when the population size is smaller.
}
\references{
Mielke, James H., Konigsberg, Lyle W., and Relethford, John H. (2011)
\emph{Human biological variation}, 2nd ed.  New York: Oxford University Press.

Mielke, James H., Konigsberg, Lyle W., and Relethford, John H. (2006)
\emph{Human biological variation}, 1st ed.  New York: Oxford University Press.
}
