% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K_reg}
\alias{K_reg}
\title{Konigsberg & Jantz (2018) regression data for stature estimation}
\format{
A data frame with 4 rows and 3 columns where
the rows are for maximum femur, tibia, humerus, and radius lengths
and the columns are the regression coefficient, the regression
constant, and the standard error of estimate. The bone measurements
should be in mm. and the returned stature estimates are in mm.
}
\usage{
data(K_reg)
}
\description{
These regression parameters are obtained by classical calibration
using the multivariate summary statistics from Konigsberg and Jantz (2018).
}
\references{
Konigsberg, L. W., & Meadows Jantz, L. (2018). Multivariate regression
 methods for the analysis of stature. In K. Latham, E. Bartelink,
 & M. Finnegan (Eds.), \emph{New Prespectives in Forensic Human Skeletal
 Identification} (pp. 87-104). Cambridge, MA: Academic Press.
}
\keyword{datasets}
