% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ped2.R
\name{make_ped2}
\alias{make_ped2}
\title{Plot pedigree from Game of Thrones and calculate inbreeding coefficents}
\usage{
make_ped2()
}
\value{
Returns pedigree and inbreeding coefficients.  We will see
in the future that the inbreeding coefficient is the probability of
identity by descent (i.b.d.), the probability that the alleles are
the same at a locus because they were inherited from the same ancestor(s).
}
\description{
Family tree for "Jon II," the presumed son of Jon Snow (Stark) and
Daenerys Targaryn.
}
\details{
OK, I got this a few years back from what I thought was the definitive
family tree.  I'm not so sure now about some of the links in this tree.  There
are no parameters as this is just a "demo" that is "hard-wired" to draw this
particular tree.
}
