% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_56.R
\name{P_56}
\alias{P_56}
\title{EM algorithm to estimate allele frequencies in ABO system}
\usage{
P_56()
}
\value{
Returns the allele frequencies and the expected phenotype counts under
Hardy-Weinberg equilibrium.
}
\description{
Calculates the allele frequencies in the ABO blood system
using the EM algorithm.
}
\details{
This object takes as input the counts for Type A blood,
Type B blood, Type AB blood, and Type O blood and find the allele frequencies
for the A allele (=p), the B allele (=q), and the O allele (=r).  From
these estimated allele frequencies it also gives the expected phenotype counts
using Hardy-Weinberg equilibrium.  There are no parameters as the object prompts
for the blood type counts. The suggested blood type counts are from Taylor and
Prior (1938).  The object is called "P_56" as page 56 in Mielke, Konigsberg,
and Relethford (2011) has an extremely abbreviated example using the Taylor and
Prior (1938) data.  Pages 52-53 (and Table 3.2) in the first edition (2006)
of the text give a more detailed treatment of the EM algorithm.
}
\references{
Taylor, G. L., & Prior, A. M. (1938). Blood groups in Britain II:
Distribution in the populations. \emph{Annals of Eugenics}, 8, 356-361.

Mielke, James H., Konigsberg, Lyle W., and Relethford, John H. (2011)
\emph{Human biological variation}, 2nd ed.  New York: Oxford University Press.

Mielke, James H., Konigsberg, Lyle W., and Relethford, John H. (2006)
\emph{Human biological variation}, 1st ed.  New York: Oxford University Press.
}
