% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V.2105}
\alias{V.2105}
\alias{V.2015}
\title{Variance-covariance matrix}
\format{
A 5 by 5 matrix
}
\usage{
data(V.2015)
}
\description{
This data is the variance-covariance matrix among stature and maximum femur,
tibia, humerus, and radius length (all in mm.) from Konigsberg and Jantz (2018).
}
\references{
Konigsberg, L. W., & Meadows Jantz, L. (2018). Multivariate regression
 methods for the analysis of stature. In K. Latham, E. Bartelink,
 & M. Finnegan (Eds.), \emph{New Prespectives in Forensic Human Skeletal
 Identification} (pp. 87-104). Cambridge, MA: Academic Press.
}
\keyword{datasets}
